+++ e2e/packages/use-infinite-scroll/src/handlers/test.ts
@@ -1,0 +1,8 @@
+export default {
+  name: "test",
+  priority: 10,
+  pattern: "/test",
+  func: async ({ link, state }) => {
+    Object.assign(state.source.data[link], { isTest: true });
+  },
+};