+++ e2e/support/index.d.ts
@@ -1,5 +1,4 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-// load type definitions that come with Cypress module
+// Load type definitions that come with Cypress module.
 /// <reference types="cypress" />
 
 declare namespace Cypress {
@@ -19,6 +18,13 @@ declare namespace Cypress {
      * ```
      */
     visitSSR(url: string): Chainable<Document>;
+
+    /**
+     * Get the HTML from a link with cy.request and validate it using amphtml-validator.
+     *
+     * @param url - The url to get the HTML from.
+     */
+    validateAMP(url: string): Chainable;
 
     /**
      * Expose custom types for cy.state().