+++ packages/html2react/processors/iframe.tsx
@@ -1,13 +1,23 @@
 import { Processor, Element } from "../types";
 import Iframe from "@frontity/components/iframe";
 
+/**
+ * Type for the Element processed by the {@link iframe} processor.
+ */
 interface IframeElement extends Element {
+  /**
+   * Props.
+   */
   props: Element["props"] & {
+    /**
+     * The optional data-src attribute.
+     */
     "data-src"?: string;
   };
 }
 
 const iframe: Processor<IframeElement> = {
+  name: "iframe",
   test: ({ node }) => node.component === "iframe",
   priority: 20,
   processor: ({ node }) => {