+++ packages/wp-source/src/actions.ts
@@ -1,6 +1,6 @@
 import { error, batch } from "frontity";
 import WpSource from "../types";
-import { parse, normalize, concatLink } from "./libraries/route-utils";
+import { addFinalSlash, concatLink } from "./libraries/route-utils";
 import { wpOrg, wpCom } from "./libraries/patterns";
 import { getMatch } from "./libraries/get-match";
 import {
@@ -15,17 +15,22 @@ import {
   is404Redirection,
   isEagerRedirection,
   fetchRedirection,
+  shouldBailRedirecting,
 } from "./utils";
 
 const actions: WpSource["actions"]["source"] = {
   fetch: ({ state, libraries }) => async (...params) => {
-    const [route, options] = params;
+    const [resource, options] = params;
     const { source } = state;
 
-    // Get route and route params.
-    const link = normalize(route);
-    const linkParams = parse(route);
-    const { query, page } = linkParams;
+    // Get the normalize and parse from libraries instead of importing them.
+    // This way they can be e.g. overriden at runtime by another package
+    const { normalize, parse, stringify } = libraries.source;
+
+    // Get link and link params.
+    const link = normalize(resource);
+    const linkParams = parse(resource);
+    const { query, page, hash } = linkParams;
 
     // Get options.
     const force = options ? options.force : false;
@@ -64,26 +69,45 @@ const actions: WpSource["actions"]["source"] = {
     // Get and execute the corresponding handler based on path.
     try {
       let { route } = linkParams;
-      // Transform route if there is some redirection.
       const redirection = getMatch(
         { route, link },
         libraries.source.redirections
       );
-      if (redirection) route = redirection.func(redirection.params);
+
+      // Create a variable that holds the value of the link that we will later
+      // pass to the `getMatch()` function. If there is a redirection, we update
+      // the value of `route`, so we have to derive the link again based on the
+      // "new" (redirected) route.
+      let linkForHandler = link;
+      if (redirection) {
+        // Transform route if there is some redirection.
+        route = redirection.func(redirection.params);
+
+        // Derive the link from the "redirected" route.
+        // We have to add the route, page, query and hash back by calling `stringify()`
+        // because they might have been removed by the redirection.
+        linkForHandler = addFinalSlash(stringify({ route, page, query, hash }));
+      }
 
       // Check if we need to check if it is a 30X redirection before fetching
       // the backend.
       if (isEagerRedirection(state.source.redirections, link)) {
         const redirection = await fetchRedirection({ link, state });
         // If there is a redirection, populate the data object and finish here.
-        if (redirection?.isRedirection) {
+        if (
+          redirection?.isRedirection &&
+          !shouldBailRedirecting(state.frontity.url + link, redirection)
+        ) {
           batch(() => Object.assign(source.data[link], redirection));
           return;
         }
       }
 
       // Get the handler for this route.
-      const handler = getMatch({ route, link }, libraries.source.handlers);
+      const handler = getMatch(
+        { route, link: linkForHandler },
+        libraries.source.handlers
+      );
 
       // Return a 404 error if no handler has matched.
       if (!handler)
@@ -133,7 +157,10 @@ const actions: WpSource["actions"]["source"] = {
       if (error.status === 404 && is404Redirection(state.source.redirections)) {
         const redirection = await fetchRedirection({ link, state });
         // If there is a redirection, populate the data object and finish here.
-        if (redirection?.isRedirection) {
+        if (
+          redirection?.isRedirection &&
+          !shouldBailRedirecting(state.frontity.url + link, redirection)
+        ) {
           batch(() => Object.assign(source.data[link], redirection));
           return;
         }