+++ packages/comscore-analytics/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @frontity/comscore-analytics
 
+## 1.3.0
+
+### Minor Changes
+
+- [`52640dfc`](https://github.com/frontity/frontity/commit/52640dfcfa51a7845329645d8b38e4cdd852b1ea) [#737](https://github.com/frontity/frontity/pull/737) Thanks [@DAreRodz](https://github.com/DAreRodz)! - Make the ComScore analytics package AMP-aware.
+
+### Patch Changes
+
+- Updated dependencies [[`4a179f5d`](https://github.com/frontity/frontity/commit/4a179f5d57cbb8c0008779e84db2b28b4017addf), [`52640dfc`](https://github.com/frontity/frontity/commit/52640dfcfa51a7845329645d8b38e4cdd852b1ea)]:
+  - frontity@1.14.3
+  - @frontity/analytics@1.4.0
+
 ## 1.2.1
 
 ### Patch Changes
@@ -68,6 +80,14 @@
 - Updated dependencies [[`3e2c355a`](https://github.com/frontity/frontity/commit/3e2c355a7530a2cda7ad74e0410389690ea57012), [`ba13f70a`](https://github.com/frontity/frontity/commit/ba13f70ae2a4360ca21c77aed1c920c02e9d45b8), [`4879e0e7`](https://github.com/frontity/frontity/commit/4879e0e7b9069c0fe2a93e02281704683616ae17), [`62fce1e5`](https://github.com/frontity/frontity/commit/62fce1e5c117faeb5902dc0ddae3b13d95cd925b)]:
   - @frontity/analytics@1.1.1
   - frontity@1.9.0
+
+## 0.1.1-infinite-scroll-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - frontity@1.8.1-infinite-scroll-beta.0
+  - @frontity/analytics@1.1.1-infinite-scroll-beta.0
 
 ## 0.1.0